Fixed a few spacing and linebreak issues
Ellipses (...) were counted as a single token and were being preceded by
a space.

Quotation and opening characters were not being added to the stack if
they were the first token.

'replace' diffs could contain a linebreak anywhere in the matched delta, causing unwanted linebreaks in the middle of swapped words.